requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
spacy
dateparser
python-dateutil
tzdata
